Updates to English
For translation purposes, I've copies the source english directory to create two new directories.  One is a directory of the high priority text. This is text that absolutely has to be very well done.  The normal still needs to be done pretty well but isn't quite as in your face as the others.