Update FlavorText.xml
I found a number of instances where the translator messed up with the code part of the file. Most frequently where [BR] or [BR][BR] appears. I also saw some instances where the code was translated for example [PLAYERFACTION1] was translated [JUGADORFACCION1]. I fixed those that I could by comparing side by side with the English file, but you will have to check this. Also, there must be many that i missed, because my concentration was on the Spanish text.