Major pass on French flavor text
This is designed to make it easier in the future.  There is now a FlavorText.fr.xliff file in the French root which we now modify and then export out to the French files.   We can import into it as well. Though we should, going forward, focus on the FlavorText_fr.xliff file.