Giant update and merge with community chnages
Thank you for the help on these updates.  We've merged the changes as best we can in with only certain changes (like removal of faction from text) but otherwise have tried to carefully pick and choose the translations that work well with the game.  If we made any mistakes we apologize. It's a very big project! :)